+++ irksome/backends/firedrake.py
@@ -7,6 +7,18 @@ from functools import reduce
 from operator import mul
 
 
+assemble = firedrake.assemble
+derivative = firedrake.derivative
+norm = firedrake.norm
+Function = firedrake.Function
+TestFunction = firedrake.TestFunction
+TrialFunction = firedrake.TrialFunction
+DirichletBC = firedrake.DirichletBC
+EquationBC = firedrake.bcs.EquationBC
+EquationBCSplit = firedrake.bcs.EquationBCSplit
+Constant = firedrake.Constant
+
+
 def get_stage_space(V: ufl.FunctionSpace, num_stages: int) -> ufl.FunctionSpace:
     # NOTE: Should use firedrake.MixedFunctionSpace, but there is a bug for vector spaces.
     return reduce(mul, (V for _ in range(num_stages)))
@@ -74,13 +86,54 @@ def invalidate_jacobian(solver):
     return firedrake.LinearVariationalSolver.invalidate_jacobian(solver)
 
 
-assemble = firedrake.assemble
-derivative = firedrake.derivative
-norm = firedrake.norm
-Function = firedrake.Function
-TestFunction = firedrake.TestFunction
-TrialFunction = firedrake.TrialFunction
-DirichletBC = firedrake.DirichletBC
-EquationBC = firedrake.bcs.EquationBC
-EquationBCSplit = firedrake.bcs.EquationBCSplit
-Constant = firedrake.Constant
+class BoundsConstrainedDirichletBC(firedrake.DirichletBC):
+    def __init__(
+        self,
+        V,
+        g,
+        sub_domain,
+        bounds,
+        solver_parameters=None,
+    ):
+        self.g = g
+        self.solver_parameters = solver_parameters
+        self.bounds = bounds
+        self.gnew = Function(V)
+
+        F = ufl.inner(self.gnew - g, TestFunction(V)) * ufl.dx
+
+        if solver_parameters is None:
+            solver_parameters = {
+                "snes_type": "vinewtonrsls",
+                "snes_max_it": 300,
+                "snes_atol": 1.0e-8,
+                "ksp_type": "preonly",
+                "mat_type": "aij",
+            }
+        problem = create_variational_problem(F, self.gnew)
+        self.solver = create_variational_solver(
+            problem, solver_parameters=solver_parameters
+        )
+        super().__init__(V, g, sub_domain)
+
+    @property
+    def function_arg(self):
+        """The value of this boundary condition."""
+        self.solver.solve(bounds=self.bounds)
+        return self.gnew
+
+    @function_arg.setter
+    def function_arg(self, g):
+        """Set the value of this boundary condition."""
+        self.solver.solve(bounds=self.bounds)
+        return self.gnew
+
+    def reconstruct(self, V=None, g=None, sub_domain=None):
+        V = V or self.function_space()
+        g = g or self.g
+        sub_domain = sub_domain or self.sub_domain
+        return type(self)(V, g, sub_domain, self.bounds, self.solver_parameters)
+
+
+def create_bounds_constrained_bc(V, g, sub_domain, bounds, solver_parameters=None):
+    return BoundsConstrainedDirichletBC(V, g, sub_domain, bounds, solver_parameters=solver_parameters)
